backend/requirements.txt
fastapi
uvicorn
numpy
Pillow
onnxruntime
python-multipart